Deprecate `CoinGeckoClient.create()` in favor of `CoinGeckoClient()` constructor
Removes the backing interface and renames the implementation to `CoinGeckoClient`.
